Reorganize to resolve circular imports
